% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_Forecast_Tails.R
\name{BC_Forecasts_Tails}
\alias{BC_Forecasts_Tails}
\title{Bias-Correction for operational forecasts with extreme tails}
\usage{
BC_Forecasts_Tails(
  o.h,
  m.h,
  m.f,
  ratio = FALSE,
  trace = 0.05,
  trace.calc = 0.5 * trace,
  jitter.factor = 0,
  n.tau = NULL,
  ratio.max = 2,
  ratio.max.trace = 10 * trace,
  ECBC = FALSE,
  ties = "first",
  subsample = NULL,
  pp.type = 7,
  upper.tail = T
)
}
\arguments{
\item{o.h}{Vector of observed values during a historical period}

\item{m.h}{Vector of modelled values during a hindcast period}

\item{m.f}{Vector of forecast values during a forecast period (1-month to 6-month ahead forecasts)}

\item{ratio}{logical value indicating if samples are of a ratio quantity. TRUE: Preserving relative ratio (e.g., precipitation), FALSE: Preserving relative absolute value (e.g., temperature)}

\item{trace}{Replace values less than trace with exact zeros (defaults to 0.05)}

\item{trace.calc}{Treat values below trace.calc as censored (defaults to 0.5*trace)}

\item{jitter.factor}{jitter to accommodate ties (defaults to 0.01). If jitter.factor > 0, apply a small amount of jitter to accommodate ties due to limited measurement precision.}

\item{n.tau}{NULL --> number of empirical quantiles (NULL=sample length)}

\item{ratio.max}{numeric value indicating the maximum proportional change allowed for ratio variables (defaults to 2)}

\item{ratio.max.trace}{Values below which ratio.max is applied (default 10*trace)}

\item{ECBC}{logical value indicating whether mhat.p outputs should be ordered according to o.h ranks using the empirical copula-bias correction (ECBC) algorithm (defaults to FALSE)}

\item{ties}{Method used to handle ties when calculating ordinal ranks (defaults to 'first')}

\item{subsample}{use subsample draws of size n.tau to calculate empirical quantiles; if NULL, calculate normally}

\item{pp.type}{type of plotting position used in quantile (defaults to 7)}

\item{upper.tail}{logical value to identify extreme values in the upper tail (e.g., precipitation and wind (minimum bound is zero) = T, temperature=F)}
}
\value{
mhat.h Vector of bias corrected m.h values for the historical period

mhat.p Vector of bias corrected m.f values for the forecast period
}
\description{
Quantile Mapping (QM) based bias-Correction considering sub-seasonal or seasonal forecasts with extremes out of historical range
}
\details{
While QM is a non-parametric approach, when forecasted extreme events are out of historical range, those extremes are fitted to an extreme probabilistic distribution (GEV) to estimate non-exceedance probability.
Thus, this is a hybrid approach with non-parametric and parametric approaches.This approach is to bias-correct for subseasonal or seasonal forecasts.
}
