% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_gev_tau.R
\name{get_gev_tau}
\alias{get_gev_tau}
\title{Returns the probability for extreme values}
\usage{
get_gev_tau(y_val, sample_data, upper = T)
}
\arguments{
\item{y_val}{Single or Array (or matrix) of extreme values beyond a historical range}

\item{sample_data}{Array(or matrix) of sample data (observed or modelled)}

\item{upper}{logical value to indicate the location of extreme values (tails on the right side or both sides). A variable with zero bound (e.g., precipitation) is upper=T.}
}
\value{
tau_extrapolated: Estimated probability for the given extreme values
}
\description{
Fits GEV to the upper or lower tail and returns the probability for extreme values out of a historical range
}
\details{
By integrating a GEV tail fit, quantiles beyond the range of a historical data can be estimated.
The GEV distribution is defined by three parameters: location, scale, and shape.
The shape parameter is particularly important as it determines the behavior of the tail.
}
