% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_gev_value.R
\name{get_gev_value}
\alias{get_gev_value}
\title{Inverse GEV}
\usage{
get_gev_value(tau, sample_data, upper = T)
}
\arguments{
\item{tau}{Single or Array (or matrix) of probability for extreme values beyond a historical range}

\item{sample_data}{Array(or matrix) of sample data (observed or modelled)}

\item{upper}{logical value to indicate the location of extreme values (tails on the right side or both sides). A variable with zero bound (e.g., precipitation) is upper=T.}
}
\value{
Inverse GEV values for given probability
}
\description{
Returns value for a given probability
}
\details{
By integrating a GEV tail fit, quantiles beyond the range of a historical data can be estimated.
For a given probability, the GEV distribution is defined with a sample data and return inverse values.
}
