% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBCDS.R
\name{MBCDS}
\alias{MBCDS}
\title{Multivariate Bias Correction with Distribution-free Shuffle approach (MBCDS)}
\usage{
MBCDS(
  o.h,
  m.h,
  m.f,
  ratio.seq = rep(FALSE, ncol(o.h)),
  trace = 0.05,
  trace.calc = 0.01,
  jitter.factor = 0,
  n.tau = NULL,
  ratio.max = 2,
  ratio.max.trace = 10 * trace,
  ties = "first",
  subsample = NULL,
  pp.type = 7,
  pr.i = 1
)
}
\arguments{
\item{o.h}{Array(or matrix) of observed values during a historical period}

\item{m.h}{Array(or matrix) of modelled values during a historical period}

\item{m.f}{Array(or matrix) of modelled values during a future (projection) period}

\item{ratio.seq}{Array of logical values indicating if samples are of a ratio quantity. TRUE: Preserving relative ratio (e.g., precipitation), FALSE: Preserving relative absolute value (e.g., temperature)}

\item{trace}{Replace values less than trace with exact zeros (defaults to 0.05)}

\item{trace.calc}{Treat values below trace.calc as censored (defaults to 0.5*trace)}

\item{jitter.factor}{jitter to accommodate ties (defaults to 0.01). If jitter.factor > 0, apply a small amount of jitter to accommodate ties due to limited measurement precision.}

\item{n.tau}{NULL --> number of empirical quantiles (NULL=sample length)}

\item{ratio.max}{numeric value indicating the maximum proportional change allowed for ratio variables (defaults to 2)}

\item{ratio.max.trace}{Values below which ratio.max is applied (default 10*trace)}

\item{ties}{Method used to handle ties when calculating ordinal ranks (defaults to 'first')}

\item{subsample}{use subsample draws of size n.tau to calculate empirical quantiles; if NULL, calculate normally}

\item{pp.type}{type of plotting position used in quantile (defaults to 7)}

\item{pr.i}{Index of precipitation variable}
}
\value{
mhat.h Array of bias corrected m.h values for the historical period

mhat.p Array of bias corrected m.p values for the projection period
}
\description{
MBCDS bias correction for preserving relative changes in quantiles and interdependence between climate variables
}
\details{
QDM is applied to bias-corrrect GCM outputs and Distribution-free Shuffle is used for preserving interdependence between climate variables.
The struture of arguments are exactly same as QDM except for multi-variables (array or matrix) of o.h, m.h, and m.p while they are vectors (univariate).
}
\references{
MBCDS:
Eum, H.-I., Gupta, A., Dibike, Y., 2020. Effects of univariate and multivariate statistical downscaling methods on climatic and hydrologic indicators for Alberta, Canada. Journal of Hydrology 588, 125065. https://doi.org/10.1016/j.jhydrol.2020.125065

QDM:
Cannon, A.J., Sobie, S.R., and Murdock, T.Q. 2015. Bias correction of simulated precipitation by quantile mapping: How well do methods preserve relative changes in quantiles and extremes? Journal of Climate, 28: 6938-6959. doi:10.1175/JCLI-D-14-00754.1
}
