% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDM.R
\name{QDM}
\alias{QDM}
\title{Quantile Delta Mapping (QDM)}
\usage{
QDM(
  o.h,
  m.h,
  m.p,
  ratio = FALSE,
  trace = 0.05,
  trace.calc = 0.5 * trace,
  jitter.factor = 0,
  n.tau = NULL,
  ratio.max = 2,
  ratio.max.trace = 10 * trace,
  ECBC = FALSE,
  ties = "first",
  subsample = NULL,
  pp.type = 7
)
}
\arguments{
\item{o.h}{Vector of observed values during a historical period}

\item{m.h}{Vector of modelled values during a historical period}

\item{m.p}{Vector of modelled values during a projection period}

\item{ratio}{logical value indicating if samples are of a ratio quantity. TRUE: Preserving relative ratio (e.g., precipitation), FALSE: Preserving relative absolute value (e.g., temperature)}

\item{trace}{Replace values less than trace with exact zeros (defaults to 0.05)}

\item{trace.calc}{Treat values below trace.calc as censored (defaults to 0.5*trace)}

\item{jitter.factor}{jitter to accommodate ties (defaults to 0.01). If jitter.factor > 0, apply a small amount of jitter to accommodate ties due to limited measurement precision.}

\item{n.tau}{NULL --> number of empirical quantiles (NULL=sample length)}

\item{ratio.max}{numeric value indicating the maximum proportional change allowed for ratio variables (defaults to 2)}

\item{ratio.max.trace}{Values below which ratio.max is applied (default 10*trace)}

\item{ECBC}{logical value indicating whether mhat.p outputs should be ordered according to o.h ranks using the empirical copula-bias correction (ECBC) algorithm (defaults to FALSE)}

\item{ties}{Method used to handle ties when calculating ordinal ranks (defaults to 'first')}

\item{subsample}{use subsample draws of size n.tau to calculate empirical quantiles; if NULL, calculate normally}

\item{pp.type}{type of plotting position used in quantile (defaults to 7)}
}
\value{
mhat.h Vector of bias corrected m.h values for the historical period

mhat.p Vector of bias corrected m.p values for the projection period
}
\description{
QDM bias correction for preserving relative changes in quantiles
}
\details{
QDM is equivalent to the equidistant and equiratio forms of quantile mapping (Cannon et al., 2015)
}
\references{
Cannon, A.J., Sobie, S.R., and Murdock, T.Q. 2015. Bias correction of
simulated precipitation by quantile mapping: How well do methods preserve
relative changes in quantiles and extremes? Journal of Climate, 28: 6938-6959. doi:10.1175/JCLI-D-14-00754.1
}
